Update : Event

+++ src/main/java/fr/evercraft/everstats/service/event/ManagerEvent.java
@@ -19,13 +19,12 @@ package fr.evercraft.everstats.service.event;
 import java.util.Optional;
 import java.util.UUID;
 
-import javax.annotation.Nullable;
-
 import org.spongepowered.api.entity.Entity;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.entity.damage.DamageType;
 
+import fr.evercraft.everapi.event.ESpongeEventFactory;
 import fr.evercraft.everapi.server.player.EPlayer;
 import fr.evercraft.everstats.EverStats;
 
@@ -38,15 +37,15 @@ public class ManagerEvent {
 	
 	public void reload() {
 		this.plugin.getLogger().debug("Event StatsReloadEvent");
-		this.plugin.getGame().getEventManager().post(new EReloadStatsSystemEvent(Cause.source(this.plugin).build()));
+		this.plugin.getGame().getEventManager().post(ESpongeEventFactory.createStatsSystemEventReload(Cause.source(this.plugin).build()));
 	}
 	
-	public void post(UUID uuid, Entity killer, DamageType cause, Long time) {
+	public void post(UUID uuid, Optional<Entity> killer, DamageType cause, Long time) {
 		Optional<EPlayer> player = this.plugin.getEServer().getEPlayer(uuid);
 		if (player.isPresent()) {
 			this.death(player.get(), killer, cause, time);
-			if (killer instanceof Player) {
-				Optional<EPlayer> killer_player = this.plugin.getEServer().getEPlayer(uuid);
+			if (killer.isPresent() && killer.get() instanceof Player) {
+				Optional<EPlayer> killer_player = this.plugin.getEServer().getEPlayer(killer.get().getUniqueId());
 				if (player.isPresent()) {
 					this.kill(player.get(), killer_player.get(), cause, time);
 				}
@@ -54,18 +53,17 @@ public class ManagerEvent {
 		}
 	}
 	
-	private void death(EPlayer victim, @Nullable Entity killer, DamageType damage, Long time) {
-		if (killer == null) {
-			this.plugin.getLogger().debug("Event StatsUserEvent.Death : (victim='" + victim.getUniqueId() + "';killer='null';damage='" + damage.getId() + "';time='" + time + "')");
-			this.plugin.getGame().getEventManager().post(new EDeathStatsUserEvent(victim, damage, time, Cause.source(this.plugin).build()));
+	private void death(EPlayer victim, Optional<Entity> killer, DamageType damage, Long time) {
+		if (killer.isPresent()) {
+			this.plugin.getLogger().debug("Event StatsUserEvent.Death : (victim='" + victim.getUniqueId() + "';killer='" + killer.get().getUniqueId() + "';damage='" + damage.getId() + "';time='" + time + "')");
 		} else {
-			this.plugin.getLogger().debug("Event StatsUserEvent.Death : (victim='" + victim.getUniqueId() + "';killer='" + killer.getUniqueId() + "';damage='" + damage.getId() + "';time='" + time + "')");
-			this.plugin.getGame().getEventManager().post(new EDeathStatsUserEvent(victim, killer, damage, time, Cause.source(this.plugin).build()));
+			this.plugin.getLogger().debug("Event StatsUserEvent.Death : (victim='" + victim.getUniqueId() + "';killer='null';damage='" + damage.getId() + "';time='" + time + "')");
 		}
+		this.plugin.getGame().getEventManager().post(ESpongeEventFactory.createStatsUserEventDeath(victim, time, damage, killer, Cause.source(this.plugin).build()));
 	}
 
 	private void kill(EPlayer victim, EPlayer killer, DamageType damage, Long time) {
 		this.plugin.getLogger().debug("Event StatsUserEvent.kill : (victim='" + victim.getUniqueId() + "';killer='" + killer.getUniqueId() + "';damage='" + damage.getId() + "';time='" + time + "')");
-		this.plugin.getGame().getEventManager().post(new EKillStatsUserEvent(victim, killer, damage, time, Cause.source(this.plugin).build()));
+		this.plugin.getGame().getEventManager().post(ESpongeEventFactory.createStatsUserEventKill(victim, time, damage, killer, Cause.source(this.plugin).build()));
 	}
 }
